messaging_app/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
drf-nested-routers==0.94.2
sqlparse==0.5.3
